Refactor DraggableQuestion
