+++ src/pages/CreateEditTestPage/CreateEditTestPage.jsx
@@ -6,10 +6,15 @@ import {
   getQuestionsSelector,
 } from 'models/test/selectors';
 import useAction from 'hooks/useAction';
-import { deleteTest } from 'models/test/reducer';
+import {
+  deleteTest,
+  setDragAndDropArrayAnswers,
+  setDragAndDropArrayQuests,
+} from 'models/test/reducer';
 import FooterTest from 'pages/CreateEditTestPage/FooterTest';
 import TestTitle from 'pages/CreateEditTestPage/TestTitle';
 import DraggableQuestion from 'pages/CreateEditTestPage/DraggableQuestion';
+import { DragDropContext, Droppable } from 'react-beautiful-dnd';
 import S from './CreateEditTestPage.styled';
 
 const CreateEditTestPage = () => {
@@ -18,10 +23,39 @@ const CreateEditTestPage = () => {
   const [uniqId, setUniqId] = useState(nanoid());
   const removeTest = useAction(deleteTest);
   const questionsEntities = useSelector(getQuestionsSelector);
+  const setDndIds = useAction(setDragAndDropArrayQuests);
+  const setDNDNewIds = useAction(setDragAndDropArrayAnswers);
 
   useEffect(() => {
     return () => removeTest();
   }, []);
+
+  const onDragEnd = result => {
+    const { destination, source, draggableId } = result;
+    if (!destination) return;
+
+    if (
+      destination.droppableId === source.droppableId &&
+      destination.index === source.index
+    ) {
+      return;
+    }
+
+    if (destination.droppableId !== source.droppableId) return;
+
+    if (destination.droppableId === 'questDrop') {
+      const copyIds = [...questionsIds];
+      copyIds.splice(source.index, 1);
+      copyIds.splice(destination.index, 0, draggableId);
+      setDndIds(copyIds);
+    } else {
+      const quests = questionsEntities[source.droppableId];
+      const copyIds = [...quests.answer.ids];
+      copyIds.splice(source.index, 1);
+      copyIds.splice(destination.index, 0, draggableId);
+      setDNDNewIds({ id: destination.droppableId, ids: copyIds });
+    }
+  };
 
   useEffect(() => {
     scrollPageToBottomTest.current.scrollIntoView({
@@ -34,10 +68,23 @@ const CreateEditTestPage = () => {
     <S.PageTest>
       <S.Content ref={scrollPageToBottomTest}>
         <TestTitle />
-        <DraggableQuestion
-          questionsIds={questionsIds}
-          questionsEntities={questionsEntities}
-        />
+        <DragDropContext onDragEnd={onDragEnd}>
+          <Droppable droppableId="questDrop" type="questItem">
+            {provided => (
+              <S.DragZone ref={provided.innerRef} {...provided.droppableProps}>
+                {questionsIds.map((id, index) => (
+                  <DraggableQuestion
+                    id={id}
+                    key={id}
+                    quest={questionsEntities[id]}
+                    index={index}
+                  />
+                ))}
+                {provided.placeholder}
+              </S.DragZone>
+            )}
+          </Droppable>
+        </DragDropContext>
         <FooterTest setUniqId={setUniqId} uniqId={uniqId} />
       </S.Content>
     </S.PageTest>
